refactor(test): split coordinator worktree tests into separate file
Move 4 worktree/cwd resolution tests from test_coordinator.py (650 lines,
exceeds 600-line exception) to new test_coordinator_worktree.py (160 lines).

This keeps test_coordinator.py focused on dispatch flow (capacity, dedup,
lifecycle, error handling) and isolates worktree/cwd resolution tests.
